+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 0.2.0 - 2016-02-19
 ### Added
 - Basic `status` command to get info about the changelog
 - Will update package.json version number too if found